Add strict_provenance and replace to_bits with expose_addr
Unstable feature strict_provenace is used. This is used to require
more strict pointer operations. The case is explained in

https://github.com/rust-lang/rust/issues/95228

"If you ever want to treat something as a Real Pointer that can be
Offset and Dereferenced, there must be an unbroken chain of custody
from that pointer to the original allocation you are trying to access
using only pointer->pointer operations. If at any point you turn a
pointer into an integer, that integer cannot be turned back into a
pointer. This includes usize as ptr, transmute, type punning with raw
pointer reads/writes, whatever. Just assume the memory "knows" it
contains a pointer and that writing to it as a non-pointer makes it
forget."

This commit includes replacement of to_bits() calls with calls to
expose_addr().

Signed-off-by: Markku Kylänpää <markku.kylanpaa@vtt.fi>