Upgrade workspace resolver setting
The rust-sdk workspace had a resolver setting `resolver = "1"` despite
workspace members have settings `resolver = "2"`. This is because the
edition 2021 in workspace.package implies `resolver = "2"` and workspace
members inherited that setting.

This was a bit confusing as also workspace.package block specified the
edition 2021 but there was still a warning. Setting the resolver
explicitly in the workspace block fixed this issue.

More information in:
https://doc.rust-lang.org/cargo/reference/resolver.html#resolver-versions

Signed-off-by: Markku Kylänpää <markku.kylanpaa@vtt.fi>