Include loader component to the builder
Builder now has a third parameter called loader in addition to
the enclave application binary and runtime binary. This is based
on Keystone modification where the code has been simplified by
using Keystone runtime loader code text section to load Keystone
runtime and an enclave binary. Rust SDK assumes that the loader
binary is still an ELF file and not just a text section. The text
section is then extracted during the loading phase.

Binaries are now specified using a Loadable type and the binaries
are stored into HashMap instead of Vec so that the host application
do not need to care about the order in which the binaries have been
added in the builder. Previously there was a bool flag that worked
with two binaries but as we now have three binaries the Loadable
type is used.

The default sizes of memory regions have been increased as well.

Signed-off-by: Markku Kylänpää <markku.kylanpaa@vtt.fi>